// Admin/src/app/services/models/campus-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TrainingCenter } from '../models/training-center';
export interface CampusResponse {
  capacity?: number;
  name?: string;
  quarter?: string;
  town?: string;
  trainingCenterCampus?: TrainingCenter;
  xcoor?: number;
  ycoor?: number;
}